projects/sms-bomber/requirements.txt
requests
stem
PySocks
